refactor: cleaning
