--- cognite/powerops/resync/config_classes/core/_core.py
@@ -1,3 +1,0 @@
-from __future__ import annotations
-
-ExternalId = str
